Feature EPL-1746: Improved the structure of context and hooks
