+++ packages/MainUI/src/contexts/languageContext.tsx
@@ -1,34 +1,8 @@
-import React, { createContext, useState, useContext, ReactNode } from 'react';
-
-type Language = 'es' | 'en';
-
-interface LanguageContextType {
-  language: Language;
-  setLanguage: (lang: Language) => void;
-}
+import { createContext } from 'react';
+import { LanguageContextType } from './types';
 
 const LanguageContext = createContext<LanguageContextType | undefined>(
   undefined,
 );
-
-export const LanguageProvider: React.FC<{ children: ReactNode }> = ({
-  children,
-}) => {
-  const [language, setLanguage] = useState<Language>('es');
-
-  return (
-    <LanguageContext.Provider value={{ language, setLanguage }}>
-      {children}
-    </LanguageContext.Provider>
-  );
-};
-
-export const useLanguage = () => {
-  const context = useContext(LanguageContext);
-  if (context === undefined) {
-    throw new Error('useLanguage must be used within a LanguageProvider');
-  }
-  return context;
-};
 
 export default LanguageContext;
