+++ packages/MainUI/src/contexts/languageProvider.tsx
@@ -1,0 +1,15 @@
+import { ReactNode, useState } from 'react';
+import LanguageContext from './languageContext';
+import { Language } from './types';
+
+export const LanguageProvider: React.FC<{ children: ReactNode }> = ({
+  children,
+}) => {
+  const [language, setLanguage] = useState<Language>('en');
+
+  return (
+    <LanguageContext.Provider value={{ language, setLanguage }}>
+      {children}
+    </LanguageContext.Provider>
+  );
+};
