+++ packages/MainUI/src/contexts/types.ts
@@ -1,0 +1,6 @@
+export type Language = 'es' | 'en';
+
+export interface LanguageContextType {
+  language: Language;
+  setLanguage: (lang: Language) => void;
+}
